requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
httpx==0.25.2
pandas==2.1.4
chardet==5.2.0
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
